song playtime int

+++ SoPra/src/main/java/org/model/Song.java
@@ -91,7 +91,7 @@ public class Song implements Comparable{
 		return title;
 	}
 
-	public double getPlaytime() {
+	public int getPlaytime() {
 		return playtime;
 	}
